pyproject.toml
[project]
name = "cra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-groq>=0.3.2",
    "langchain-ollama>=0.3.3",
    "mcp[cli]>=1.9.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]
